+++ Assets/Scripts/ParentVillage/Events/SendChildToWellEventScript.cs
@@ -23,7 +23,7 @@ public class SendChildToWellEventScript : InteractableBuildingEventScript
     
     protected override bool YesButtonEnabledImpl { get { return true; } }
     protected override string NoButtonTextImpl { get { return "No"; } }
-    public override float CostToPerform { get { return 0; } }
+    public override int CostToPerform { get { return 0; } }
     protected override float LockTime { get { return TimeManager.SecondsPerYear / 6.0f; } }
 
     public override BuildingType BuildingType { get { return BuildingType.Well; } }
